Fixed error 4

+++ lb2/lab2main.cc
@@ -27,9 +27,9 @@ void image_invert(const uint8_t* rgb, uint8_t* inverted, int num_pixels) {
 	int j = 0;
 	auto t1 = chrono::high_resolution_clock::now();
 	for(int i = 0; i < num_pixels; ++i, rgb+=3) {
-		int ch1 = ~rgb[0];
-		int ch2 = ~rgb[1];
-		int ch3 = ~rgb[2];
+		uint8_t ch1 = ~rgb[0];
+		uint8_t ch2 = ~rgb[1];
+		uint8_t ch3 = ~rgb[2];
 		inverted + j = ch1;
 		j+=8;
 		inverted + j = ch2;
